# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Application for planning your journey",
        "title": "Journer Planner",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/signIn": {
            "post": {
                "description": "Authorization to the account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "SignIn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "your login or email",
                        "name": "option",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "your password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/singUp": {
            "post": {
                "description": "Create account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "SignUp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "your login",
                        "name": "login",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "your password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "your email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/tasks/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "AddTask",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Details",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-10-21",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "14:00",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "name": "days",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 2,
                        "name": "hours",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 30,
                        "name": "minutes",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/tasks/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete existing task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "DeleteTask",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id of group",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/tasks/getlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "GetTasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/tasks/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update existing task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "UpdateTask",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Details",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-10-21",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "14:00",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "name": "days",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 2,
                        "name": "hours",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 30,
                        "name": "minutes",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}